requirements.txt
Django
djangorestframework
psycopg
# psycopg2-binary
# pillow
python-dotenv
django-filter
djangorestframework-simplejwt
drf_yasg
django-cors-headers
redis
celery
django_celery_beat
django_filter
requests
flake8

#Windows
eventlet
